requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
python-dotenv==1.0.1

numpy==2.1.3
scikit-learn==1.5.2
networkx==3.4.2
PyYAML==6.0.2

pytest==8.3.4

streamlit==1.41.1
requests==2.32.3
pandas==2.2.3
PyYAML==6.0.2